tg-bot/requirements.txt
python-telegram-bot==21.1.1
peewee==3.17.3
psycopg2-binary